requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
django-cors-headers==4.9.0
django-ipware==7.0.1
django-ratelimit==4.1.0
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
ipinfo==5.2.1
joblib==1.5.2
kombu==5.5.4
multidict==6.7.0
numpy==2.3.4
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-ipware==3.0.0
redis==7.0.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
